+++ src/writer/Writer.java
@@ -3,5 +3,5 @@ package writer;
 import java.io.IOException;
 
 public interface Writer {
-	void write(int[] sequence, long count) throws IOException, InterruptedException;
+	void write(int[] sequence, double count) throws IOException, InterruptedException;
 }
